Add a query cache and keep crawl reports in memory
Crawling a spec takes time (typically a few seconds). To make sure the service
only crawls specs when needed, a new cache was added at the query level for
POST requests that send the HTML contents of the specs to crawl.

Note this cache is restricted to POST requests that contain the HTML of the
specs to crawl for now: in the absence of HTML content, it is hard to tell
whether the crawl result we have in cache is still valid (the HTML content of
the spec may have changed in the meantime) without re-fetching the spec.

Cache entries are periodically refreshed (once per day by default), to account
for indirect updates that do not touch on the HTML content itself (e.g. updates
to scripts that the spec depends on).

Main crawl reports (from reffy-reports) are now also kept in memory to avoid
having to parse a lot of JSON at each request. This saves ~70ms per request.